kindle-requirements.txt
airtable-python-wrapper
ipython